changes package name and spelling errors
